Support local:// raster tile URLs
In the style JSON configuration, allow to use the local://
pseudo-protocol prefix also for raster tile source URLs. When serving
the style JSON, rewrite it into an absolute URL using the public URL or
the Host header from the request.

This allows to use raster tile sources which are served from the same
domain, but externally from tileserver-gl. The requests could be handled
by a reverse proxy in front of tileserver-gl for example, and redirected
to a different service or static files.

Instead of hardcoding an absolute host name in the style configuration,
using a relative URL allows to use the same style configuration on
different hosts and on localhost in the development environment.