fix blocking of tileserver while serving index
we noticed sporadic long blocks of the entier tileserver on version 3.1.1
They always happend when somebody accessed the frontPage.
Our guess is, that clone() blocks the object and takes very long to finish.

We replaced the clone with ES6 magic from https://stackoverflow.com/a/42988763
and the problem is gone.

Also the time to generate the index goesm from ~3s to ~0.2s
This should primarily affect user with many large styles